requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.28
pillow==10.2.0
pip==23.3.2
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4
wheel==0.42.0
whitenoise==6.6.0


gunicorn